+++ LtiLibrary.AspNetCore.Tests/Lis/v2/MembershipControllerShould.cs
@@ -6,6 +6,7 @@ using LtiLibrary.AspNetCore.Tests.SimpleHelpers;
 using Xunit;
 using LtiLibrary.NetCore.Lis.v2;
 using LtiLibrary.NetCore.Lti1;
+using LtiLibrary.NetCore.Membership;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.PlatformAbstractions;
